% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMLFreeModel.R
\name{fitMLFreeModel}
\alias{fitMLFreeModel}
\title{Fit the MLE for the free model.}
\usage{
fitMLFreeModel(data, phy, init = NULL, root.type = "madfitz",
  bounds = c(0, 100), opts = NULL, n.cores = 1)
}
\arguments{
\item{data}{}

\item{phy}{}

\item{init}{}

\item{root.type}{}

\item{bounds}{}

\item{opts}{}

\item{n.cores}{}
}
\value{
A list with the fit, solution and other information.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
Here each of the sites in the data is fit a different MK model. So this is the most parameter rich model.
}
\author{
daniel
}
