% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMLGammaModel.R
\name{fitMLGammaModel}
\alias{fitMLGammaModel}
\title{Find the maximum likelihood for the model with the gamma function.}
\usage{
fitMLGammaModel(data, phy, root.type = "madfitz", ncat = 4, bounds = c(0,
  100), opts = NULL, n.cores = 1)
}
\arguments{
\item{data}{matrix with species names as rownames.}

\item{phy}{phylogeny}

\item{root.type}{one of "madfitz", or "equal". Need to extend to accept a observed vector of probabilities.}

\item{ncat}{categories for the gamma function}

\item{bounds}{}

\item{opts}{the list of options for nloptr. If null it will use the default parameters.}

\item{n.cores}{}

\item{bonds}{a numeric vector with the lower and upper bonds for the rates}
}
\value{
A list with the log-likelihood, initial parameters and the parameter values.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
daniel
}
