% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMLGammaSpanSpace.R
\name{fitMLGammaSpanSpace}
\alias{fitMLGammaSpanSpace}
\title{Find the maximum likelihood for the model with the gamma function.}
\usage{
fitMLGammaSpanSpace(data, phy, root.type = "madfitz", ncat = 4,
  bounds = c(0, 100), state.space = "max.div", opts = NULL, init = NULL,
  verbose = TRUE, n.cores = 1)
}
\arguments{
\item{data}{matrix with species names as rownames.}

\item{phy}{phylogeny}

\item{root.type}{one of "madfitz", or "equal". Need to extend to accept a observed vector of probabilities.}

\item{ncat}{categories for the gamma function}

\item{bounds}{}

\item{state.space}{can be "max.div", "site.div" or a numeric. The size of the state space is equal to the largest observed or equal to the per site state diversity plus the number.}

\item{opts}{the list of options for nloptr. If null it will use the default parameters.}

\item{init}{}

\item{verbose}{}

\item{n.cores}{}

\item{bonds}{a numeric vector with the lower and upper bonds for the rates}
}
\value{
A list with the log-likelihood, initial parameters and the parameter values.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
daniel
}
