% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMLStateDiversity.R
\name{fitMLStateDiversity}
\alias{fitMLStateDiversity}
\title{Find the maximum likelihood for the state diversity model.}
\usage{
fitMLStateDiversity(data, phy, rate.model = "gamma", symmetric = TRUE,
  root.type = "madfitz", add.invar = FALSE, ncat = 4, bounds = NULL,
  div.max = "observed", opts = NULL, init = NULL, verbose = TRUE,
  n.cores = 1)
}
\arguments{
\item{data}{matrix with species names as rownames.}

\item{phy}{phylogeny}

\item{rate.model}{options are "correlated", "gamma", and "single.rate"}

\item{symmetric}{if the model have equal rates for birth and death events.}

\item{root.type}{one of "madfitz", or "equal". Need to extend to accept a observed vector of probabilities.}

\item{add.invar}{TRUE or FALSE. Whether to compute transition rates for the invariant sites or not.}

\item{ncat}{categories for the gamma function}

\item{bounds}{a 2x2 numeric matrix. First column are the limits for the birth rate and second column are limits for the death rate. First line is the lower limit and second line is the upper limit.}

\item{div.max}{can be "observed" for the limits given the observed diversity on the position or a numeric value with the number of states to be added above and below the observed max and min diversity in each position.}

\item{opts}{the list of options for nloptr. If null it will use default parameters.}

\item{init}{parameters to initialize the search}

\item{verbose}{whether to print messages to the screen.}

\item{n.cores}{number of cores to split the computation of the likelihood}
}
\value{
A list with the log-likelihood, initial parameters and the parameter values.
}
\description{
Fits the state diversity model.
}
\details{
This model is different from the sequence trait model in the sense that it does not focus on rates of transitions between states on each of the positions of a sequence trait. This models make inferences about the rate in which the diversity of a given sequence position increase or decrease compared to the other rates across the sequence trait.
}
\author{
daniel
}
