% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAverageRates.R
\name{getAverageRates}
\alias{getAverageRates}
\title{Model average rates along the positions of the sequence trait}
\usage{
getAverageRates(..., verbose = TRUE, legacy.mode = FALSE)
}
\arguments{
\item{...}{a list of model fit or a series of model fits separated by commas.}

\item{verbose}{whether to print information to the screen.}

\item{legacy.mode}{whether to check structure of objects instead of class.}
}
\value{
A list object with a vector of rates, a vector of flags from the model averaging algorithm and a named vector with the AICw for the models.
}
\description{
Function performs model averaging given the AIC weights of the models. This follows a similar procedure to the one implemented in Caetano et al., 2018.
}
\details{
Model averaging is performed by multiplying the rate estimated under each of the models by the respective model weight. Model weight is a measure of the relative likelihood of a model. This can be used to associate a relative likelihood to parameter estimates. See more information on Caetano et al., 2018.
}
\author{
Daniel Caetano
}
