% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogLikMk.R
\name{getLogLikMk}
\alias{getLogLikMk}
\title{Compute the loglik for the model.}
\usage{
getLogLikMk(phy, data, Q, root.type)
}
\arguments{
\item{phy}{the phylogeny}

\item{data}{the data matrix}

\item{Q}{the transition matrix}

\item{root.type}{the root type}
}
\value{
Return the loglik for the model.
}
\description{
Internal function to get the likelihood of the model.
}
\details{
No details.
}
\author{
daniel
}
