% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRootMarginal.R
\name{getRootMarginal}
\alias{getRootMarginal}
\title{Function to compute the marginal estimate of root values.}
\usage{
getRootMarginal(phy, data, fit)
}
\arguments{
\item{phy}{}

\item{data}{}

\item{fit}{}
}
\value{
A list with the marginal root probabilities.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
daniel
}
