% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLogLik.R
\name{makeLogLik}
\alias{makeLogLik}
\title{Likelihood function for the models}
\usage{
makeLogLik(
  data,
  phy,
  se = NULL,
  model.type,
  rate.model,
  ncat = NULL,
  n.cores = 1,
  verbose = TRUE
)
}
\arguments{
\item{data}{matrix with species names as rownames and sequence positions as the columns.}

\item{phy}{ultrametric phylogeny in ape 'phylo' format}

\item{se}{the standard error of the tip data. This should be a named vector with length equal to tip data with the measurement error for each species. Alternatively, this can be a single value that will be used for all species. Set to NULL if unknown.}

\item{model.type}{the type of model for the likelihood function. Options are "MK_discrete" for data with discrete states and discrete positions or "BM_discrete" for data with continuous states and discrete positions.}

\item{rate.model}{options are "correlated" , "gamma", and "single.rate". See Details.}

\item{ncat}{number of categories for the discrete Gamma distribution}

\item{n.cores}{number of cores used for parallel computation of the likelihood.}

\item{verbose}{if informative messages should be printed.}
}
\value{
a likelihood function.
}
\description{
Likelihood function for the models
}
\details{
This is a general likelihood maker that will return a likelihood function for different models implemented in the package. The rate.model controls the type of rate models. The options are "correlated" for models using the bi-variate Discrete Gamma model (auto-discrete-gamma, Yang 1995), "gamma" for models using the independent Discrete Gamma distribution, and "single-rate" for models in which there is a single rate for all sequence positions.
}
