% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFasta2Matrix.R
\name{readFasta2Matrix}
\alias{readFasta2Matrix}
\title{Read a FASTA file and returns a matrix}
\usage{
readFasta2Matrix(file)
}
\arguments{
\item{file}{the FASTA file to be read.}
}
\value{
a matrix
}
\description{
Reads a FASTA file of any type and returns a matrix with the species as rows and the sites as columns.
}
\details{
Each site of the FASTA file is transformed into a "character" in the matrix. The function ony works for sequences of the same length (i.e., alignments).
}
\author{
daniel
}
