% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMCMC.R
\name{readMCMC}
\alias{readMCMC}
\title{Read the MCMC from files.}
\usage{
readMCMC(file.name, burn = 0.5, thin = 10)
}
\arguments{
\item{file.name}{}

\item{burn}{}

\item{thin}{}
}
\value{
A data matrix with the samples.q
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
daniel
}
