% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMCFreeModel.R
\name{runMCMCFreeModel}
\alias{runMCMCFreeModel}
\title{A simple MCMC sampler for the model.}
\usage{
runMCMCFreeModel(phy, data, root.type = "equal", gen, unif.prior = c(0,
  100), init = NULL)
}
\arguments{
\item{phy}{}

\item{data}{}

\item{root.type}{}

\item{gen}{}

\item{unif.prior}{}
}
\value{
Write posterior samples to the directory.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
daniel
}
