% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMCMCSimple.R
\name{runMCMCSimple}
\alias{runMCMCSimple}
\title{A simple MCMC sampler for the model.}
\usage{
runMCMCSimple(phy, data, k = 4, gen, Q.max = 100, beta.max = "estimated")
}
\arguments{
\item{phy}{}

\item{data}{}

\item{k}{}

\item{gen}{}

\item{Q.max}{}

\item{beta.max}{}
}
\value{
Write posterior samples to the directory.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
daniel
}
