% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqtraits.R
\docType{package}
\name{seqtraits}
\alias{seqtraits}
\alias{seqtraits-package}
\title{seqtraits}
\description{
Estimate rates of evolution for sequence traits.
}
\author{
Daniel Caetano <caetanods1@gmail.com>
}
