% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMLGammaBM.R
\name{fitCorrSeqBM}
\alias{fitCorrSeqBM}
\title{Maximum likelihood estimate of evolutionary rates for sequence phenotype with continuous states and discrete positions}
\usage{
fitCorrSeqBM(
  data,
  phy,
  se = NULL,
  rate.model = "gamma",
  ncat = 4,
  init.M = FALSE,
  bounds = NULL,
  opts = NULL,
  search.global = FALSE,
  init = NULL,
  verbose = TRUE,
  n.cores = 1
)
}
\arguments{
\item{data}{matrix with species names as rownames and sequence positions as the columns. Each sequence position should be a numeric value, within the bounds of the "BM_bounds" matrix.}

\item{phy}{ultrametric phylogeny in ape 'phylo' format}

\item{se}{the standard error of the tip data. This should be a named vector with length equal to tip data with the measurement error for each species. Alternatively, this can be a single value that will be used for all species. Set to NULL if unknown.}

\item{rate.model}{options are "correlated" , "gamma", and "single.rate". See Details below.}

\item{ncat}{categories for the gamma function.}

\item{init.M}{whether the initial state for the M matrix (for the correlated model) should have starting state equal to the gamma model (i.e., equal probabilities for all the transitions) or a random starting point.}

\item{bounds}{a numeric vector of length 2 with the lower and upper bonds for the BM rates.}

\item{opts}{the list of options for nloptr. If NULL it will use the default parameters of this function (not the same as the default for 'nloptr'). See more information in the help page for 'nloptr'.}

\item{search.global}{whether to perform a global MLE search before the local MLE search. Default is FALSE.}

\item{init}{set the initial parameters for the MLE search. The length varies depending on the rate.model .}

\item{verbose}{whether to print information to the screen.}

\item{n.cores}{number of cores to perform the likelihood evaluation.}
}
\value{
A list with the log-likelihood, initial parameters and the parameter values.
}
\description{
Function uses the discrete-gamma model of Yang (1994) to estimate variation in rates of evolution independent of sequence position and the auto-discrete-gamma model of Yang (1995) when rates are correlated among sites. This function takes a sequence trait with continuous states and discrete positions. The approach fits a bounded Brownian motion model to the data, with bounds defined by the user. Bounds can be estimated, but that is not recommended. Each sequence position can have their own bound.
}
\author{
Daniel Caetano
}
