% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogLikMk.R
\name{getLogLikMk}
\alias{getLogLikMk}
\title{Compute the loglik for the model.}
\usage{
getLogLikMk(phy, data, Q, root.type)
}
\arguments{
\item{phy}{}

\item{data}{}

\item{Q}{}

\item{root.type}{}
}
\value{
Return the loglik for the model.
}
\description{
.. content for \description{} (no empty lines) ..
}
\details{
.. content for \details{} ..
}
\author{
daniel
}
