% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRates.R
\name{getRates}
\alias{getRates}
\title{Extract rates of evolution for sequence positions.}
\usage{
getRates(x, Q.model = "ER", gap.char = "-")
}
\arguments{
\item{x}{a model fit result.}

\item{Q.model}{one of "ER" or "DEL"}

\item{gap.char}{the char used to identify gaps.}
}
\value{
a vector or a matrix
}
\description{
Extracts and summarize the estimates of rates of evolution from the fitted models.
}
\details{
Returns a vector if the model is single rate and a matrix if the model is including different rates for the transitions between states and between the gap states. This functions will substitute invariant positions (which rates were not estimated) for NA.
}
\author{
daniel
}
