% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRootMarginal.R
\name{getRootMarginal}
\alias{getRootMarginal}
\title{Function to compute the marginal estimate of root values.}
\usage{
getRootMarginal(phy, data, fit)
}
\arguments{
\item{phy}{the phylogeny}

\item{data}{the data matrix}

\item{fit}{the model fit}
}
\value{
A list with the marginal root probabilities.
}
\description{
Get the marginal probability at the root.
}
\details{
No details.
}
\author{
daniel
}
